version checking complete

+++ db/migrate/20250530212327_create_access_data.rb
@@ -1,7 +1,7 @@
 class CreateAccessData < ActiveRecord::Migration[7.1]
   def change
     create_table :access_data do |t|
-      t.integer :database_version
+      t.decimal :database_version
 
       t.timestamps
     end
